よく考えたらDataHandlerはクラスである必要が無かったので別ファイルに移動
